chore: reconcile securedrop-updater reorganization
Adapted from
freedomofpress/securedrop-updater@cf81d3277e0f7e5318c6b1c07ebdc36c8c7af1c0c3320d92591e0bf402347623.
A notable compromise is moving moving sdw_{notify,updater,util} into the
top of the package so that "python setup.py install" installs them
directly into site-packages.  We keep the existing paths under
"./launcher" in place as symlinks so that the virtual environments and
their requirements remain separate for testing.
