+++ MANIFEST.in
@@ -4,8 +4,6 @@ include dom0/*.j2
 include dom0/*.yml
 include dom0/*.conf
 include dom0/remove-tags
-include dom0/securedrop-login
-include dom0/securedrop-launcher.desktop
 include dom0/securedrop-handle-upgrade
 include dom0/update-xfce-settings
 include README.md
@@ -15,9 +13,8 @@ include sd-proxy/*
 include sd-whonix/*
 include sd-app/*
 include sd-workstation/*
-include launcher/*.py
-include launcher/sdw_updater_gui/*.py
-include launcher/sdw_notify/*.py
-include launcher/sdw_util/*.py
+include sdw_updater/*.py
+include sdw_notify/*.py
+include sdw_util/*.py
 include usb-autoattach/*
 include files/*
