+++ files/sdw-updater.desktop
@@ -1,0 +1,7 @@
+[Desktop Entry]
+Version=1.0
+Type=Application
+Terminal=false
+Icon=securedrop
+Name=SecureDrop
+Exec=sdw-updater
