+++ launcher/sdw_notify
@@ -1,0 +1,1 @@
+../sdw_notify
