+++ launcher/sdw_updater
@@ -1,0 +1,1 @@
+../sdw_updater
