+++ launcher/tests/test_updater.py
@@ -8,7 +8,7 @@ from tempfile import TemporaryDirectory
 from unittest import mock
 from unittest.mock import call
 
-relpath_updater_script = "../sdw_updater_gui/Updater.py"
+relpath_updater_script = "../sdw_updater/Updater.py"
 path_to_script = os.path.join(os.path.dirname(os.path.abspath(__file__)), relpath_updater_script)
 updater = SourceFileLoader("Updater", path_to_script).load_module()
 from Updater import UpdateStatus  # noqa: E402
@@ -792,7 +792,7 @@ def test_apply_dom0_state_success(mocked_info, mocked_error, mocked_subprocess):
 def test_apply_dom0_state_failure(mocked_info, mocked_error, mocked_subprocess):
     updater.apply_dom0_state()
     log_error_calls = [
-        call("Failed to apply dom0 state. See launcher-detail.log for details."),
+        call("Failed to apply dom0 state. See updater-detail.log for details."),
         call("Command 'check_output' returned non-zero exit status 1."),
     ]
     mocked_subprocess.assert_called_once_with(
