+++ sdw_notify/strings.py
