+++ sdw_updater/strings.py
